--- Manero-backend/DTOs/User/RegisterDto.cs
@@ -1,11 +1,0 @@
-﻿namespace Manero_backend.DTOs.User
-{
-    public class RegisterDto
-    {
-        public string Username { get; set; } = null!;
-        public string Password { get; set; } = null!;
-        public string Email { get; set; } = null!;
-        public string FirstName { get; set; } = null!;
-        public string LastName { get; set; } = null!;
-    }
-}
