+++ Manero-backend/DTOs/User/SignInDto.cs
@@ -1,0 +1,19 @@
+﻿using Microsoft.AspNetCore.Identity;
+using System.ComponentModel.DataAnnotations;
+
+namespace Manero_backend.DTOs.User
+{
+    public class LoginDto
+    {
+        [Required]
+        public string Email { get; set; } = null!;
+
+        [Required]
+        public string Password { get; set; } = null!;
+
+        public string RememberMe { get; set; } = null!;
+        public string ForgotPassword { get; set; } = null!;
+        
+      
+    }
+}
