+++ Manero-backend/DTOs/User/SignUpDto.cs
@@ -1,0 +1,25 @@
+﻿using System.ComponentModel.DataAnnotations;
+
+namespace Manero_backend.DTOs.User
+{
+    public class RegisterDto
+    {
+        [Required]
+        public string FirstName { get; set; } = null!;
+
+        [Required]
+        public string LastName { get; set; } = null!;
+
+        [Required]
+        public string Email { get; set; } = null!;
+
+        [Required]
+        public string Password { get; set; } = null!;
+
+        [Required]
+        public string ConfirmPassword { get; set; } = null!;
+         
+        
+    }
+}
+
