--- Manero-backend/DTOs/User/UserDto.cs
@@ -1,10 +1,0 @@
-﻿namespace Manero_backend.DTOs.User
-{
-    public class UserDto
-    {
-        public string Username { get; set; } = null!;
-        public string Email { get; set; } = null!;
-        public string FirstName { get; set; } = null!;
-        public string LastName { get; set; } = null!;
-    }
-}
